requirements.txt
openai==0.27.8
torch==2.3.0
numpy==1.25.2
pandas==2.0.3
tensorflow==2.15.0
keras==2.15.0
python-dotenv==1.0.1
transformers==4.31.0
pyyaml==6.0.1
pytest==7.4.4
geopy==2.3.0
Nominatim==0.1
tenacity==8.3.0
jinja2==3.1.4
typer==0.9.4
gradio==3.36.1
requests==2.32.3
tqdm==4.66.4
faiss-cpu==1.7.4
